Conversions of current to totes to speed!
Ever been perplexed by the sheer amount of numbers required to convert
current to totes, and totes to speed? With this package of VIs, you
don't have to!  These VIs are so well integrated, that they do the
conversions for you.  You don't need to worry about all of those
confusing numbers anymore, with current-to-totes and totes-to-speed.
WARNING: Excessive Levels of Arbitrariness, exercise caution when
operating VIs
-Nicolas Welty 2015