New global variable in AutoStack
Hi, um, our marketing and advertising team was on holiday, so some of
the engineers needed to type up this ad instead.  This variable is
inserted into the AutoStack, and it will return whether or not the
encoder has reached the tolerances and the Soft Limit set in the code
(Which we mean is the VARS.txt file).  My sincere apologies for this
rather blatant interference, and we will try much harder to align their
vacation days away from official "focal sale" points.
-with sincere apology,
Nicolas Welty, Senior Engineer and Code Dev. 2015