The Next Commit! Interruptions that WORK!
The next stage is here.  We now have INTERRUPTS.  Some call this magic,
but we call this the next step in code-based engineering.  No longer
will you have to use the WAIT function, because now you can WAIT until a
sensor trips, which will prove to be exceedingly useful.  This has been
implemented into the code, and must be re-purchased.  If you bring your
old Access Codes, we will give you a 4-5% discount!
-Nicolas Welty 2015